requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.6.0
-e git+https://github.com/wjs2063/fastapi_backend.git@4a5ac1c5bfe8661df1c1429e7693f4699dd3308c#egg=app
attrs==25.4.0
bcrypt==4.3.0
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cssselect==1.2.0
cssutils==2.11.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email-validator==2.3.0
emails==0.6
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
frozenlist==1.8.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.28.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-core==1.2.2
langchain-openai==1.1.5
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.3
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.7.0
mypy==1.11.2
mypy-extensions==1.0.0
neo4j==6.1.0
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.0.1
nodeenv==1.9.1
openai==2.13.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
premailer==3.10.0
propcache==0.4.1
psycopg==3.2.2
psycopg-binary==3.2.2
psycopg-pool==3.3.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.18.0
PyJWT==2.10.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
ruff==0.6.7
sentry-sdk==1.45.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.24
starlette==0.38.6
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typer==0.12.5
types-passlib==1.7.7.20240819
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.3
uuid_utils==0.12.0
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.5
watchfiles==0.24.0
websockets==13.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
